pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperclaw"
version = "1.0.7"
description = "Unlimited-domain multi-agent AI swarm platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0",
    "asyncpg>=0.29",
    "pgvector>=0.2",
    "httpx>=0.27",
    "typer>=0.12",
    "rich>=13.0",
    "textual>=0.61",
    "anthropic>=0.28",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[project.scripts]
hyperclaw = "cli.hyperclaw:main"
hyperclaw-tui = "hyperclaw.tui:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hyperclaw*", "core*", "memory*", "models*", "swarm*", "security*", "ui*", "recursive*", "cli*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
